+++ src/NodemcuTools.ts
@@ -67,6 +67,12 @@ export default class NodemcuTools {
 	public async connect(devicePath: string): Promise<INodeMcu> {
 		const device = NodeMcuRepository.getOrCreate(devicePath)
 		await device.connect()
+		const delay = workspace
+			.getConfiguration()
+			.get('nodemcu-tools.connectionDelay', initialSettings.connectionDelay)
+		if (delay && delay > 0) {
+			await device.delayConnection(delay)
+		}
 		await device.detectEspType()
 
 		await commands.executeCommand('setContext', 'nodemcu-tools:isConnected', true)